chore: Cleaning up profile tests per feedback in #6553
