+++ docs/src/content/docs/05-community/01-contributing.mdx
@@ -219,7 +219,13 @@ The largest group of opt-in tests are the ones that drive a real OpenTofu or Ter
 go test -tags=tf -run '^TestTF' ./...
 ```
 
-Without that tag, `go test ./...` only runs the tests that Terragrunt can satisfy on its own, so the suite stays green in an environment with no OpenTofu or Terraform installed. Add the tag whenever you touch behavior that reaches the binary.
+Tests that depend on OpenTofu specifically, rather than on whichever binary `tf-path` points at, are prefixed with `TestTofu*` and tagged with the `tofu` build tag. Run them alongside the `TestTF*` tests, the way CI does:
+
+```bash
+go test -tags=tofu,tf -run '^(TestTF|TestTofu)' ./...
+```
+
+Without those tags, `go test ./...` only runs the tests that Terragrunt can satisfy on its own, so the suite stays green in an environment with no OpenTofu or Terraform installed. Add the tag whenever you touch behavior that reaches the binary.
 
 Terragrunt also includes integration tests for Google Cloud Platform (GCP). These tests are prefixed with `TestGcp*` and are tagged with the `gcp` build tag. To run these tests, you can use the `-tags` flag set in the `GOFLAGS` environment variable, similar to AWS tests:
